# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/cinema_halls": {
            "get": {
                "tags": [
                    "CinemaHall"
                ],
                "summary": "Get a listing of the CinemaHalls.",
                "description": "Get all CinemaHalls",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/CinemaHall"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CinemaHall"
                ],
                "summary": "Store a newly created CinemaHall in storage",
                "description": "Store CinemaHall",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "CinemaHall that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/CinemaHall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CinemaHall"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/cinema_halls/{id}": {
            "get": {
                "tags": [
                    "CinemaHall"
                ],
                "summary": "Display the specified CinemaHall",
                "description": "Get CinemaHall",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CinemaHall",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CinemaHall"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CinemaHall"
                ],
                "summary": "Update the specified CinemaHall in storage",
                "description": "Update CinemaHall",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CinemaHall",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "CinemaHall that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/CinemaHall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CinemaHall"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CinemaHall"
                ],
                "summary": "Remove the specified CinemaHall from storage",
                "description": "Delete CinemaHall",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of CinemaHall",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CinemaHall": {
            "required": [
                "name",
                "totalSeat"
            ],
            "properties": {
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "totalSeat": {
                    "description": "totalSeat",
                    "type": "integer",
                    "format": "int32"
                },
                "cinema_id": {
                    "description": "cinema_id",
                    "type": "integer",
                    "format": "int32"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}